Added levantine cultures
Added Hebrew culture into: European set and all-world set.

Changed namebase of Carthaginian in Antique set from Berber (17) to Levantine (42) as per the conversation on discord thread FMG suggestions > Punic culture set.
This leaves the antique set with no native speaker of berber. If you want to have a culture that represent north-african people and berber different than carthage, we can add one.

Included Phoenician as a distinct culture with similar culture center formula than carthage "sort: i => n(i) / td(i, 20) / sf(i)" but with td(i, 20) because is the temperature of beirut and gives a difference with carthaginian.

Hebrew is included with " n(i) / td(i, 18) * sf(i) " because hebrew are pastoralists, and to differentiate the culture center from the phoenician that are near sea harbor.